moved snoflake utils
